Bin errors taken into account in checks
The comparson with the check threshols is now performed taking into account
the error on the ratio, with a tolerance of +/- 2 sigma by default.
The corresponding checkDeviationNsigma parameter has been added to the
muon plots configurations.
The configuration for the global muon tracking has been split into one
for MCH-MID and one for MFT-MCH-MID.