Fixed averaging of ratio plots
The averaging of plots to produce the denominators for the check was
not correctly handling the case where the plots are actually of type
TH1Ratio. The issue is fixed by creating a new TH1D for the first plot
in the averaging process, instead of cloning the existing TH1Ratio.